refactor(usage-index): decompose parseClaude/parseCodex/aggregate and split into usage-parsers.mjs + usage-aggregate.mjs
parseClaude (CC 33), parseCodex (CC 53), and aggregate (CC 44) each
decompose into small per-record/per-phase helper functions, all under
CC 25. usage-index.mjs (1,736 raw / 1,067 counted lines) then splits
along its natural seams:

- usage-parsers.mjs: the per-vendor transcript parsers (parseClaude,
  parseCodex, blankSession, addUsage, projectLabel, IDLE_GAP_MS) and
  their private helpers.
- usage-aggregate.mjs: pure arithmetic over already-parsed records
  (mergeIntervals, maskSecrets, normalizeSessionIdentity,
  applyCodexLedger, aggregate, sessionPayload, MAX_TURN_CHARS) — a leaf
  module with no dependency on its siblings.
- usage-index.mjs: index I/O, the on-disk cache, build/scan
  orchestration, and single-session resolution; re-exports the moved
  names so every existing import path (tests, dashboard-server,
  usage-opencode.mjs) keeps working unchanged.

usage-opencode.mjs now imports blankSession/addUsage directly from
usage-parsers.mjs instead of usage-index.mjs, removing the pre-existing
circular import between the two (usage-index.mjs -> usage-opencode.mjs
-> usage-parsers.mjs is now a straight line, not a cycle).

All three files land under the 1,000-counted-line max-lines threshold
(usage-index.mjs 771 raw lines / usage-parsers.mjs 594 / usage-aggregate.mjs
517), and every eslint complexity warning in this file family is gone.
No behavior change: identical aggregation output, identical cache
schema, identical exports. Realigns every usage-index.mjs citation in
docs/USAGE-SCORECARD-METRICS.md and docs/TRANSCRIPTS.md that moved
file or line.
