+++ src/lib/usage-index.mjs
@@ -28,9 +28,15 @@
 // pricing/usage-classify/usage-insights are loaded LAZILY and are injectable as
 // `deps`. That seam is what lets the scanner be unit-tested against exact
 // arithmetic without importing anyone's pricing table or classification policy.
+//
+// This file owns index I/O (file discovery, the on-disk cache, single-session
+// resolution) and the build/scan orchestration. The per-vendor transcript
+// parsers live in usage-parsers.mjs; the pure aggregation/session-shape
+// arithmetic lives in usage-aggregate.mjs. Both are re-exported below where a
+// consumer's existing import path expects them from here.
 import fs from 'node:fs';
 import path from 'node:path';
-import { configDir, claudeDir, codexDir, repoRoot } from './paths.mjs';
+import { configDir, claudeDir, codexDir } from './paths.mjs';
 import { readCodexStateResult } from './codex-state.mjs';
 import {
   defaultOpencodeDbPath, listSessionsResult as listOpencodeSessionsResult,
@@ -40,7 +46,11 @@ import {
   addTelemetryDiagnostics, emptyTelemetryDiagnostics, finalizeTelemetryDiagnostics,
   MAX_TELEMETRY_UNKNOWN_KINDS, recordTelemetryUnit, telemetryCapabilities,
 } from './usage-telemetry.mjs';
-import { decodeClaudeRecord, decodeCodexRecord } from './telemetry-records.mjs';
+import { parseClaude, parseCodex } from './usage-parsers.mjs';
+import { maskSecrets, applyCodexLedger, aggregate, sessionPayload } from './usage-aggregate.mjs';
+
+export { IDLE_GAP_MS, projectLabel } from './usage-parsers.mjs';
+export { MAX_TURN_CHARS, mergeIntervals, maskSecrets, normalizeSessionIdentity, applyCodexLedger } from './usage-aggregate.mjs';
 
 /** Bump to invalidate every cached entry wholesale.
  *  v2: cached records carry `active` sub-intervals for the idle-gap split.
@@ -81,683 +91,16 @@ import { decodeClaudeRecord, decodeCodexRecord } from './telemetry-records.mjs';
  *       prompts/responses, so every Codex record must be re-derived. */
 export const SCHEMA_VERSION = 10;
 
-/** Silence longer than this ends a stretch of engagement. A session is split
- *  into active sub-intervals at gaps ABOVE this bound (exactly this much is not
- *  a gap), and `engagedSeconds` unions those. Named rather than inline because
- *  it is a judgement call the numbers depend on, not a magic constant. */
-export const IDLE_GAP_MS = 15 * 60 * 1000;
-
 const DAY_MS = 86_400_000;
 // One day of slack past dashboard-server.mjs's 365-day clampDays ceiling —
 // see the carry-forward pruning comment in scan() below.
 const KEEP_MS = 366 * DAY_MS;
-export const MAX_TURN_CHARS = 40_000;
 /** Largest single transcript readSession will pull into memory. The corpus's
  *  biggest real file is ~18 MB; JSON expansion runs ~5x, so 64 MB caps the
  *  spike near 320 MB instead of unbounded. Above this the session reads as
  *  unavailable rather than risking the panel's process. */
 const MAX_SESSION_BYTES = 64 * 1024 * 1024;
 const VALID_ID = /^[A-Za-z0-9._-]{1,128}$/;
-
-// ── pure helpers ────────────────────────────────────────────────────────────
-
-/** Milliseconds from an epoch number, Date, or ISO string; NaN when unusable. */
-function toMs(v) {
-  if (typeof v === 'number') return Number.isFinite(v) ? v : NaN;
-  if (v instanceof Date) return v.getTime();
-  if (typeof v === 'string') return Date.parse(v);
-  return NaN;
-}
-
-/**
- * Total seconds covered by a set of intervals, counting overlap ONCE.
- * Accepts `[start, end]` tuples or `{ start, end }` objects; each bound may be
- * an epoch number, a Date, or an ISO string. Degenerate or unparseable
- * intervals are dropped rather than throwing. Pure — exported for test.
- */
-export function mergeIntervals(intervals) {
-  if (!Array.isArray(intervals)) return 0;
-  const spans = [];
-  for (const iv of intervals) {
-    if (!iv) continue;
-    const start = toMs(Array.isArray(iv) ? iv[0] : iv.start);
-    const end = toMs(Array.isArray(iv) ? iv[1] : iv.end);
-    if (!Number.isFinite(start) || !Number.isFinite(end) || end <= start) continue;
-    spans.push([start, end]);
-  }
-  if (!spans.length) return 0;
-  spans.sort((a, b) => a[0] - b[0]);
-  let total = 0;
-  let [curStart, curEnd] = spans[0];
-  for (let i = 1; i < spans.length; i++) {
-    const [s, e] = spans[i];
-    if (s <= curEnd) {           // overlapping OR exactly touching → extend
-      if (e > curEnd) curEnd = e;
-    } else {
-      total += curEnd - curStart;
-      curStart = s; curEnd = e;
-    }
-  }
-  total += curEnd - curStart;
-  return Math.round(total / 1000);
-}
-
-// Each replacement keeps the human-readable prefix and drops the payload. The
-// replacement text cannot re-match its own pattern (the '…' is outside every
-// character class), which is what makes masking idempotent.
-/** @type {[RegExp, string][]} */
-// Order matters. PEM blocks are matched WHOLE and first: a later pattern that
-// nibbled at the base64 body would leave the armour behind, which reads as
-// "masked" on screen while the key material is still visible.
-const SECRET_PATTERNS = [
-  [/-----BEGIN [A-Z ]*PRIVATE KEY-----[\s\S]*?-----END [A-Z ]*PRIVATE KEY-----/g,
-    '-----BEGIN PRIVATE KEY----- …redacted -----END PRIVATE KEY-----'],
-  [/-----BEGIN [A-Z ]*PRIVATE KEY-----/g, '-----BEGIN PRIVATE KEY----- …redacted'],
-  [/\bsk-[A-Za-z0-9_-]{12,}/g, 'sk-…redacted'],
-  [/\bgh[pousr]_[A-Za-z0-9]{20,}/g, 'ghp_…redacted'],
-  [/\bAKIA[0-9A-Z]{12,}/g, 'AKIA…redacted'],
-  [/\bwhsec_[A-Za-z0-9_-]{16,}/g, 'whsec_…redacted'],
-  [/\bASIA[0-9A-Z]{12,}/g, 'ASIA…redacted'],          // STS temporary credential
-  [/\bgithub_pat_[A-Za-z0-9_]{20,}/g, 'github_pat_…redacted'],
-  [/\bglpat-[A-Za-z0-9_-]{16,}/g, 'glpat-…redacted'],
-  [/\bhf_[A-Za-z0-9]{20,}/g, 'hf_…redacted'],
-  [/\bpypi-[A-Za-z0-9_-]{16,}/g, 'pypi-…redacted'],
-  [/\b(?:sk|rk)_live_[A-Za-z0-9]{16,}/g, 'sk_live_…redacted'],
-  [/\bSG\.[A-Za-z0-9_-]{16,}\.[A-Za-z0-9_-]{16,}/g, 'SG.…redacted'],
-  [/\bnpm_[A-Za-z0-9]{30,}/g, 'npm_…redacted'],
-  [/\bxox[baprse]-[A-Za-z0-9-]{10,}/g, 'xox…redacted'],
-  [/\bAIza[0-9A-Za-z_-]{30,}/g, 'AIza…redacted'],
-  // Slack incoming-webhook URLs are bearer credentials in URL form.
-  [/https:\/\/hooks\.slack\.com\/services\/[A-Za-z0-9/]+/g,
-    'https://hooks.slack.com/services/…redacted'],
-  // JWT — three base64url segments. Masked whole; the payload is the sensitive part.
-  [/\beyJ[A-Za-z0-9_-]{8,}\.[A-Za-z0-9_-]{8,}\.[A-Za-z0-9_-]{8,}/g, 'eyJ…redacted'],
-  // Inline credentials in a URI (postgres://user:pass@host, https://u:p@h). The
-  // USERNAME is deliberately preserved — it is usually needed to identify which
-  // system leaked, and it is not the secret.
-  [/\b([a-z][a-z0-9+.-]*):\/\/([^\s:@/]+):[^\s@/]+@/g, '$1://$2:…redacted@'],
-  [/\b(Basic)\s+[A-Za-z0-9+/]{16,}={0,2}/g, '$1 …redacted'],
-  // Case-insensitive on the SCHEME only — `BEARER`/`bearer`/`Bearer` all appear
-  // in the wild. The token body stays case-sensitive, so this cannot widen into
-  // prose the way an /i on the assignment rule would.
-  [/\b([Bb]earer|BEARER)\s+[A-Za-z0-9._~+/=-]{16,}/g, '$1 …redacted'],
-  // Context-carried secrets: SCREAMING_CASE assignments whose NAME says secret.
-  // Deliberately case-SENSITIVE: with /i this matches prose like
-  // "tokens used = 10028979467", and these transcripts discuss token counts
-  // constantly. Uppercase-only tracks the actual env-var convention instead.
-  [/\b([A-Z][A-Z0-9_]*(?:SECRET|TOKEN|PASSWORD|PASSWD|API_?KEY|PRIVATE_KEY)[A-Z0-9_]*)(\s*[:=]\s*)("?)[^\s"']{8,}\3/g,
-    '$1$2$3…redacted$3'],
-  // Quoted JSON/JS object key whose name says secret — "apiKey": "…", 'client_secret': '…'.
-  // Case-insensitive is safe here: the quote delimiters are a shape prose never
-  // has, so this cannot widen into "tokens used" the way the assignment rule's
-  // /i would (that rule stays case-sensitive above for exactly that reason).
-  [/(["'][A-Za-z0-9_-]*(?:secret|token|password|passwd|api_?key|private_?key)[A-Za-z0-9_-]*["']\s*:\s*)(["'])[^"']{8,}\2/gi,
-    '$1$2…redacted$2'],
-  // Line-anchored YAML/TOML/ini assignment — api_key = …, password: … — with no
-  // quoting required. Anchored to line-start-through-key-through-:/= so it
-  // cannot match a key phrase floating mid-sentence ("tokens used = 10028979467"
-  // fails: "used" — not a secret-shaped word — sits directly before "=").
-  [/^([ \t]*[A-Za-z0-9_-]*(?:secret|token|password|passwd|api_?key|private_?key)[A-Za-z0-9_-]*[ \t]*[:=][ \t]*)(["']?)[^\s"']{8,}\2/gim,
-    '$1$2…redacted$2'],
-];
-
-// NOT masked, deliberately: a bare 40-char base64-ish AWS secret with no prefix
-// and no assignment context. It is indistinguishable from a hash, a checksum, or
-// a base64 blob in ordinary prose, and masking it would corrupt real transcript
-// content. It is caught when it appears as AWS_SECRET_ACCESS_KEY=… above.
-
-/**
- * Mask secret-shaped strings in transcript text (ADR-0009 §8). Idempotent, and
- * a no-op on ordinary prose. Pure — exported for test.
- */
-export function maskSecrets(text) {
-  if (text === null || text === undefined) return '';
-  let out = typeof text === 'string' ? text : String(text);
-  for (const [re, sub] of SECRET_PATTERNS) out = out.replace(re, sub);
-  return out;
-}
-
-/** Local calendar day, `YYYY-MM-DD`. Local because "what did I spend today" is
- *  a question about the user's clock, not UTC's. */
-function localDay(ms) {
-  const d = new Date(ms);
-  const p = (n) => String(n).padStart(2, '0');
-  return `${d.getFullYear()}-${p(d.getMonth() + 1)}-${p(d.getDate())}`;
-}
-
-/** `dow-hour` punchcard key, local, with Monday as 0. */
-function punchKey(ms) {
-  const d = new Date(ms);
-  return `${(d.getDay() + 6) % 7}-${d.getHours()}`;
-}
-
-function clip(text, max = 100) {
-  const t = String(text ?? '').replace(/\s+/g, ' ').trim();
-  return t.length > max ? `${t.slice(0, max - 1)}…` : t;
-}
-
-/**
- * Directory names that mean "the thing below me is a WORKTREE of the repo above
- * me", not a project of its own. `path.basename(cwd)` on a worktree yields the
- * branch/phase name — so `keel/.autopilot/worktrees/agent-runtime/phase-1`
- * reported a project called `phase-1`, and eight such rows sat beside `keel` in
- * the tree as if they were peer repositories. Every project total was wrong by
- * however much work happened in a worktree.
- */
-const WORKTREE_MARKERS = ['.autopilot', '.claude', '.git'];
-
-/**
- * Resolve a session's cwd to `{ project, worktree }`.
- *
- * The worktree name is KEPT, not discarded: "which repo" and "which branch of
- * it" are different questions, and collapsing the second into the first would
- * trade one wrong answer for a lossy one. Pure — exported for test.
- *
- * A session run in a SUB-DIRECTORY of a repo is the same problem the worktree
- * markers above solve, reached by a different path shape: `emailibrium/backend`
- * reports a project called `backend`, which then sits beside `emailibrium` as
- * if it were a peer repository. The markers cannot catch it because there is no
- * marker segment to match — only the repository boundary distinguishes them. So
- * the caller may supply `repoRoot`, and when it does, the repo becomes the
- * project and the sub-path takes the worktree slot ("which branch/part of it"),
- * exactly as a real worktree does.
- *
- * `repoRoot` is a PARAMETER rather than a lookup because resolving it is a
- * filesystem walk and this function is pure and called once per session; the
- * index resolves it once per distinct cwd and passes it in.
- *
- * @param {string|null} cwd Absolute path the session ran in.
- * @param {string|null} [dirName] Encoded ~/.claude/projects dir, used only as a
- *   fallback when the transcript carried no cwd.
- * @param {string|null} [repoRoot] The repository root containing `cwd`, when known.
- * @returns {{ project: string, worktree: string|null }}
- */
-export function projectLabel(cwd, dirName, repoRoot) {
-  if (cwd && typeof cwd === 'string') {
-    // Split on BOTH separators, not path.sep. On Windows path.sep is '\\', but a
-    // transcript's recorded cwd may be POSIX-style (WSL, a synced dotfile, a
-    // fixture) — splitting on the host separator alone yields one segment and
-    // silently disables worktree detection on that platform.
-    const segs = cwd.split(/[\\/]+/).filter(Boolean);
-
-    // <repo>/<marker>/worktrees/<...rest>  →  repo = <repo>, worktree = rest
-    for (let i = 1; i < segs.length - 1; i++) {
-      if (WORKTREE_MARKERS.includes(segs[i]) && segs[i + 1] === 'worktrees') {
-        const rest = segs.slice(i + 2).join('/');
-        return { project: segs[i - 1], worktree: rest || null };
-      }
-    }
-
-    // Claude Code's per-session scratchpad lives under the OS temp dir, not in
-    // any repo. It is genuinely not a project, so it gets its own bucket rather
-    // than a guessed one — the embedded path segment is `/`-encoded and cannot
-    // be decoded unambiguously (a `-` may be a separator or part of a name).
-    // Positional indexing is wrong here — the temp root varies (`/tmp/...` vs
-    // `/private/tmp/...`), so match the marker segment wherever it lands.
-    // `segs` already holds the trailing component, and unlike path.basename() it
-    // is separator-agnostic — basename('/a/b') is 'b' on POSIX but the whole
-    // string on Windows, which would have made every project label wrong there.
-    const base = segs[segs.length - 1];
-
-    if (segs.includes('scratchpad') && segs.some((seg) => /^claude-\d+$/.test(seg))) {
-      return { project: 'scratchpad', worktree: base || null };
-    }
-
-    // A sub-directory of a known repo is that repo, not a peer of it. Compared
-    // on split segments rather than string prefixes so `/a/repo-two` is never
-    // read as living inside `/a/repo`.
-    if (repoRoot && typeof repoRoot === 'string') {
-      const rootSegs = repoRoot.split(/[\\/]+/).filter(Boolean);
-      const inside = rootSegs.length < segs.length
-        && rootSegs.every((seg, i) => segs[i] === seg);
-      if (inside) {
-        return { project: rootSegs[rootSegs.length - 1], worktree: segs.slice(rootSegs.length).join('/') || null };
-      }
-    }
-
-    if (base && base !== '.') return { project: base, worktree: null };
-  }
-
-  if (!dirName) return { project: 'unknown', worktree: null };
-  const parts = String(dirName).replace(/^-+/, '').split('-').filter(Boolean);
-  return { project: parts.length ? parts[parts.length - 1] : 'unknown', worktree: null };
-}
-
-/** Repo root for a cwd, memoized for the life of one index build.
- *
- *  Indexing walks thousands of sessions but only tens of distinct working
- *  directories, so the filesystem walk is paid once per directory rather than
- *  once per session. A cwd outside any repo memoizes `null` — a miss is as
- *  worth caching as a hit, and `null` then leaves projectLabel on its existing
- *  basename path. */
-function repoRootMemo(resolve = repoRoot) {
-  const cache = new Map();
-  return (cwd) => {
-    if (typeof cwd !== 'string' || !cwd) return null;
-    if (!cache.has(cwd)) {
-      let root;
-      try { root = resolve(cwd); } catch { root = null; }
-      cache.set(cwd, root ?? null);
-    }
-    return cache.get(cwd);
-  };
-}
-
-/** Module-scoped because the parse functions are called per session from
- *  several entry points and threading a cache through all of them would add a
- *  parameter to each for no behavioural gain. Safe to share: a repository root
- *  does not move while a process runs, and the key space is the machine's
- *  distinct working directories (tens), not its sessions (thousands). */
-const repoRootOf = repoRootMemo();
-
-/** Sum a record's per-model usage rows into one API-equivalent cost. Rows with
- *  an observed transcript cost (opencode) use it — same preference as aggregate. */
-function sessionCost(rec, deps) {
-  let cost = 0;
-  for (const row of rec.usage ?? []) {
-    cost += row.costObserved != null ? row.costObserved : (deps.costOf({
-      model: row.model, provider: rec.provider,
-      input: row.input, output: row.output, cacheRead: row.cacheRead, cacheWrite: row.cacheWrite,
-    }) || 0);
-  }
-  return round(cost);
-}
-
-/** Write a projectLabel() result onto a record without losing the worktree. */
-function applyProject(rec, res) {
-  rec.project = res.project;
-  if (res.worktree) rec.worktree = res.worktree;
-}
-
-/**
- * Split the historical transcript-source `provider` field into an explicit host
- * and independently evidenced inference provider. The legacy field used
- * `claude`/`codex` to mean transcript host; those values therefore cannot prove
- * Anthropic/OpenAI provider identity.
- */
-export function normalizeSessionIdentity(record = {}) {
-  const legacyHost = !record.host && ['claude', 'codex'].includes(record.provider)
-    ? record.provider : null;
-  const host = record.host ?? legacyHost ?? null;
-  const provider = legacyHost ? null : (record.provider ?? null);
-  return {
-    ...record,
-    host,
-    provider,
-    model: record.model ?? null,
-    providerProvenance: provider ? (record.providerProvenance ?? 'unknown') : 'unknown',
-  };
-}
-
-// ── transcript parsing ──────────────────────────────────────────────────────
-
-/** Split JSONL into parsed objects, skipping anything that will not parse. */
-function* jsonLines(raw) {
-  for (const line of raw.split('\n')) {
-    if (!line || line.charCodeAt(0) !== 123 /* '{' */) continue;
-    let obj;
-    try { obj = JSON.parse(line); } catch { continue; }
-    if (obj && typeof obj === 'object') yield obj;
-  }
-}
-
-/** A blank per-session record; `usage` rows are (day, model) buckets so byDay
- *  and byModel can both be derived without re-reading the transcript. Exported
- *  so other transcript-source parsers (usage-opencode.mjs) build the SAME
- *  record shape instead of hand-mirroring it. */
-export function blankSession(id, provider) {
-  return {
-    id, provider, host: provider, inferenceProvider: null, providerProvenance: 'unknown',
-    title: '', project: 'unknown', start: null, end: null,
-    prompts: 0, responses: 0, exceptions: 0, sidechain: false, threadSource: null, models: [], tools: {},
-    skill: null, plugin: null, worktree: null, usage: [], punchcard: {}, active: [], stamps: [],
-    // Codex-only detail (v6): reasoning tokens inside output, and the last
-    // rate-limit snapshot the rollout carried. Claude sessions keep the zero
-    // and the null — absent, not unknown.
-    reasoningOutput: 0, rateLimits: null,
-  };
-}
-
-function noteSpan(rec, ms) {
-  if (!Number.isFinite(ms)) return;
-  if (rec.start === null || ms < rec.start) rec.start = ms;
-  if (rec.end === null || ms > rec.end) rec.end = ms;
-  rec.stamps.push(ms);
-}
-
-/**
- * Collapse a session's activity timestamps into the intervals it was actually
- * working, splitting wherever the transcript went quiet for longer than
- * IDLE_GAP_MS. A run of one timestamp yields a zero-length interval and so
- * contributes nothing — an instant has no duration to claim.
- */
-function activeIntervals(stamps) {
-  const ts = stamps.filter(Number.isFinite).sort((a, b) => a - b);
-  if (!ts.length) return [];
-  const out = [];
-  let start = ts[0];
-  let prev = ts[0];
-  for (let i = 1; i < ts.length; i++) {
-    if (ts[i] - prev > IDLE_GAP_MS) { out.push([start, prev]); start = ts[i]; }
-    prev = ts[i];
-  }
-  out.push([start, prev]);
-  return out;
-}
-
-/** Finish a parsed record: derive active intervals, drop the raw timestamps
- *  (thousands per session, and never needed again once collapsed). */
-function seal(rec) {
-  rec.active = activeIntervals(rec.stamps);
-  delete rec.stamps;
-  return rec;
-}
-
-/** Add usage to a session's (day, model) bucket, creating it on first touch.
- *  Returns the row so a caller with a per-source extra field (opencode's
- *  observed `costObserved`) can set it without a second find(). Exported for
- *  the same reason as blankSession — one definition of "how a usage row
- *  accumulates", shared across transcript-source parsers. */
-export function addUsage(rec, day, model, u) {
-  let row = rec.usage.find((r) => r.day === day && r.model === model);
-  if (!row) { row = { day, model, input: 0, output: 0, cacheRead: 0, cacheWrite: 0, responses: 0 }; rec.usage.push(row); }
-  row.input += u.input; row.output += u.output;
-  row.cacheRead += u.cacheRead; row.cacheWrite += u.cacheWrite;
-  row.responses += u.responses ?? 0;
-  return row;
-}
-
-/**
- * Harness-output envelopes: user-role entries whose text the HARNESS wrote —
- * background-task notifications, command stdout/stderr dumps, local-command
- * caveats. They carry neither `isMeta` nor a tool_result block, so text shape
- * is the only signal. Measured on the real corpus (envelope at start of user
- * text): task-notification 550, bash-stdout 85, local-command-stdout 60,
- * local-command-caveat 183; the stderr variants are the symmetric error-path
- * siblings. NOT here: bash-input (the person typed that `! cmd`) and the
- * command-name/-message/-args triple (the person invoked that slash command).
- */
-const HARNESS_OUTPUT_RE = /^\s*<(task-notification|bash-stdout|bash-stderr|local-command-stdout|local-command-stderr|local-command-caveat)>/;
-
-function entryText(entry) {
-  const content = entry?.message?.content;
-  if (typeof content === 'string') return content;
-  if (!Array.isArray(content)) return '';
-  const t = content.find((b) => b?.type === 'text' && typeof b.text === 'string');
-  return t ? t.text : '';
-}
-
-/** Is this `user` entry a human prompt, or harness output being fed back? */
-function isHumanPrompt(entry) {
-  if (entry.isMeta) return false;
-  if (HARNESS_OUTPUT_RE.test(entryText(entry))) return false;
-  const content = entry?.message?.content;
-  if (typeof content === 'string') return content.trim().length > 0;
-  if (!Array.isArray(content)) return false;
-  const hasResult = content.some((b) => b?.type === 'tool_result');
-  const hasText = content.some((b) => b?.type === 'text' && String(b.text ?? '').trim());
-  return hasText && !hasResult;
-}
-
-/**
- * What KIND of `user`-role turn is this, for transcript attribution? The
- * Messages API records tool results and harness context injections under
- * `role: "user"`, so role alone must never be read as "the human typed this":
- *   'tool-result' — carries a tool_result block: output the HARNESS fed back
- *                   to the model after a tool call.
- *   'context'     — isMeta OR a harness-output envelope (task notifications,
- *                   command stdout/stderr, caveats): harness-injected, not
- *                   typed by the person — and not the model either.
- *   'prompt'      — the human. Deliberately broader than isHumanPrompt():
- *                   an image-only paste has no text block (so it is not
- *                   COUNTED as a prompt) but it IS the person acting, and
- *                   labeling it "tool result" would misattribute it. Also
- *                   covers bash-input (`! cmd`) and slash-command records —
- *                   the person initiated those.
- */
-function userTurnKind(entry) {
-  const content = entry?.message?.content;
-  if (Array.isArray(content) && content.some((b) => b?.type === 'tool_result')) return 'tool-result';
-  if (entry.isMeta || HARNESS_OUTPUT_RE.test(entryText(entry))) return 'context';
-  return 'prompt';
-}
-
-/**
- * Parse one Claude transcript. Returns `{ session, turns }`; `turns` is only
- * populated when `withTurns` (the reader path) — the scan path does not need
- * message bodies and holding them would balloon memory over 3,000 files.
- */
-function parseClaude(raw, { id, dirName, withTurns = false }) {
-  const rec = blankSession(id, 'claude');
-  const turns = [];
-  let firstPrompt = '';
-  let aiTitle = '';
-
-  for (const e of jsonLines(raw)) {
-    const ms = toMs(e.timestamp);
-    if (e.type === 'ai-title') { if (typeof e.aiTitle === 'string') aiTitle = e.aiTitle; continue; }
-    if (typeof e.attributionSkill === 'string' && !rec.skill) rec.skill = e.attributionSkill;
-    if (typeof e.attributionPlugin === 'string' && !rec.plugin) rec.plugin = e.attributionPlugin;
-    const decoded = decodeClaudeRecord(e);
-    if (decoded.isSidechain) rec.sidechain = true;
-    if (rec.project === 'unknown' && typeof e.cwd === 'string') applyProject(rec, projectLabel(e.cwd, dirName, repoRootOf(e.cwd)));
-
-    if (decoded.role === 'user') {
-      noteSpan(rec, ms);
-      const human = isHumanPrompt(e);
-      if (human) {
-        rec.prompts++;
-        if (!firstPrompt) firstPrompt = decoded.text;
-      }
-      if (withTurns && decoded.text) {
-        turns.push({ role: 'user', at: new Date(ms).toISOString(), text: decoded.text, prompt: human, kind: userTurnKind(e) });
-      }
-      continue;
-    }
-
-    if (decoded.role !== 'assistant' || !e.message) continue;
-    noteSpan(rec, ms);
-    rec.responses++;
-    const at = Number.isFinite(ms) ? ms : (rec.start ?? Date.now());
-    const pk = punchKey(at);
-    rec.punchcard[pk] = (rec.punchcard[pk] ?? 0) + 1;
-
-    // A dropped connection, rate limit, or auth failure makes Claude Code
-    // synthesize a local placeholder turn (model: "<synthetic>",
-    // isApiErrorMessage: true) with no real completion behind it — usage is
-    // always zero. It IS real engaged time (counted above), but it is not a
-    // model attempt: excluded from `models`/cost attribution so it can never
-    // appear as a $0 "model in play," and counted instead as an EXCEPTION so
-    // it stays visible rather than silently vanishing. isApiErrorMessage isn't
-    // reliably set on every build that emits this placeholder, so the literal
-    // model marker is checked directly too — it's the one part of the shape
-    // that's never varied in observed transcripts.
-    if (decoded.isApiError) {
-      rec.exceptions++;
-      if (withTurns) {
-        turns.push({
-          role: 'assistant', at: new Date(at).toISOString(), model: 'exception',
-          text: decoded.text, tools: [], exception: true,
-        });
-      }
-      continue;
-    }
-
-    const model = typeof decoded.model === 'string' ? decoded.model : 'unknown';
-    if (!rec.models.includes(model)) rec.models.push(model);
-
-    addUsage(rec, localDay(at), model, { ...decoded.usage, responses: 1 });
-
-    const tools = [];
-    for (const use of decoded.toolUses) {
-      if (typeof use.name === 'string') {
-        tools.push(use.name);
-        rec.tools[use.name] = (rec.tools[use.name] ?? 0) + 1;
-      }
-    }
-    if (withTurns) {
-      turns.push({
-        role: 'assistant', at: new Date(at).toISOString(), model,
-        text: decoded.text, tools,
-      });
-    }
-  }
-
-  rec.title = maskSecrets(aiTitle || clip(firstPrompt)) || '(untitled)';
-  if (rec.project === 'unknown') applyProject(rec, projectLabel(null, dirName));
-  return { session: seal(rec), turns };
-}
-
-function codexParseStats() {
-  return {
-    legacyEvents: 0, itemCompletedEvents: 0, tokenCountEvents: 0,
-    prompts: 0, responses: 0, unknownItemTypes: {}, unknownItemTypeOverflow: 0,
-  };
-}
-
-function recordCodexUnknownType(stats, type) {
-  if (Object.hasOwn(stats.unknownItemTypes, type)) {
-    stats.unknownItemTypes[type]++;
-  } else if (Object.keys(stats.unknownItemTypes).length < MAX_TELEMETRY_UNKNOWN_KINDS) {
-    stats.unknownItemTypes[type] = 1;
-  } else {
-    stats.unknownItemTypeOverflow++;
-  }
-}
-
-/**
- * Parse one Codex rollout. `total_token_usage` is CUMULATIVE, so the LAST
- * token_count event is the session total — summing them would multiply the
- * figure by the number of turns. `input_tokens` there INCLUDES
- * `cached_input_tokens`, which bill as cache reads, so the two are separated.
- *
- * A rollout whose `session_meta.thread_source` is `subagent` is a delegated
- * thread whose file replays its parent thread's ENTIRE prior token history
- * as duplicate events before its own new turns (openai/codex thread_spawn
- * behavior — see ccusage/ccusage#950, which measured up to 91x cost
- * inflation from exactly this). Its cumulative `total_token_usage` therefore
- * double-counts tokens the parent session already billed, so it is excluded
- * from cost/token aggregation here; the session record itself is kept
- * (`threadSource` is surfaced on it) so it stays visible/auditable.
- */
-function parseCodex(raw, { id, withTurns = false }) {
-  const rec = blankSession(id, 'codex');
-  const turns = [];
-  const stats = codexParseStats();
-  let lastUsage = null;
-  let lastUsageAt = null;
-  let firstPrompt = '';
-
-  for (const e of jsonLines(raw)) {
-    const ms = toMs(e.timestamp);
-    noteSpan(rec, ms);
-    const decoded = decodeCodexRecord(e);
-
-    if (decoded.type === 'meta') {
-      if (typeof decoded.sessionId === 'string' && decoded.sessionId) rec.id = decoded.sessionId;
-      if (typeof decoded.cwd === 'string') applyProject(rec, projectLabel(decoded.cwd, null, repoRootOf(decoded.cwd)));
-      if (typeof decoded.threadSource === 'string') rec.threadSource = decoded.threadSource;
-      if (decoded.provider) {
-        rec.inferenceProvider = decoded.provider;
-        rec.providerProvenance = 'observed';
-      }
-      continue;
-    }
-    if (decoded.type === 'turnContext') {
-      if (typeof decoded.model === 'string' && !rec.models.includes(decoded.model)) rec.models.push(decoded.model);
-      if (decoded.provider) {
-        rec.inferenceProvider = decoded.provider;
-        rec.providerProvenance = 'observed';
-      }
-      if (rec.project === 'unknown' && typeof decoded.cwd === 'string') applyProject(rec, projectLabel(decoded.cwd, null, repoRootOf(decoded.cwd)));
-      continue;
-    }
-    if (e.type !== 'event_msg') continue;
-
-    if (decoded.type === 'tokenCount') {
-      stats.tokenCountEvents++;
-      const t = decoded.usage.total;
-      if (t) { lastUsage = t; lastUsageAt = ms; }
-      // Every token_count also carries a live rate-limit snapshot — keep the
-      // LAST one, normalized. Field names are a trap upstream: `primary` is
-      // whichever window the server listed first, NOT reliably the 5-hour one
-      // (observed live: primary = the 10080-minute weekly). So windows are
-      // kept as a flat list keyed by window_minutes and never by field name.
-      const rl = decoded.usage.rateLimits;
-      if (rl) {
-        const windows = [];
-        for (const w of [rl.primary, rl.secondary]) {
-          if (!w || typeof w !== 'object') continue;
-          const usedPercent = Number(w.used_percent);
-          const windowMinutes = Number(w.window_minutes);
-          if (!Number.isFinite(usedPercent) || !Number.isFinite(windowMinutes)) continue;
-          windows.push({
-            usedPercent, windowMinutes,
-            resetsAt: Number.isFinite(Number(w.resets_at)) ? Number(w.resets_at) : null,
-          });
-        }
-        if (windows.length) {
-          rec.rateLimits = {
-            at: Number.isFinite(ms) ? ms : null,
-            limitId: typeof rl.limit_id === 'string' ? rl.limit_id : null,
-            planType: typeof rl.plan_type === 'string' ? rl.plan_type : null,
-            windows,
-          };
-        }
-      }
-      continue;
-    }
-
-    if (decoded.generation === 'legacy') stats.legacyEvents++;
-    else if (decoded.generation === 'item') stats.itemCompletedEvents++;
-    if (decoded.unknownItemType) recordCodexUnknownType(stats, decoded.unknownItemType);
-    if (decoded.type !== 'message') continue;
-
-    if (decoded.role === 'user') {
-      rec.prompts++;
-      stats.prompts++;
-      const text = decoded.text;
-      if (!firstPrompt) firstPrompt = text;
-      // Codex rollouts record only real prompts as user_message events — tool
-      // output travels in other event types that are not surfaced as turns —
-      // so every normalized Codex user turn is kind 'prompt' by construction.
-      if (withTurns && text) turns.push({ role: 'user', at: new Date(ms).toISOString(), text, prompt: true, kind: 'prompt' });
-      continue;
-    }
-    rec.responses++;
-    stats.responses++;
-    const at = Number.isFinite(ms) ? ms : (rec.start ?? Date.now());
-    const pk = punchKey(at);
-    rec.punchcard[pk] = (rec.punchcard[pk] ?? 0) + 1;
-    if (withTurns) {
-      turns.push({
-        role: 'assistant', at: new Date(at).toISOString(),
-        model: rec.models[rec.models.length - 1] ?? 'unknown',
-        text: decoded.text, tools: [],
-      });
-    }
-  }
-
-  if (lastUsage && rec.threadSource !== 'subagent') {
-    const cacheRead = Number(lastUsage.cached_input_tokens) || 0;
-    const gross = Number(lastUsage.input_tokens) || 0;
-    const at = Number.isFinite(lastUsageAt) ? lastUsageAt : (rec.end ?? rec.start ?? Date.now());
-    addUsage(rec, localDay(at), rec.models[rec.models.length - 1] ?? 'unknown', {
-      input: Math.max(0, gross - cacheRead),
-      output: Number(lastUsage.output_tokens) || 0,
-      cacheRead,
-      cacheWrite: 0,
-      responses: rec.responses,
-    });
-    // Reasoning tokens are a SUBSET of output_tokens (they bill as output) —
-    // recorded as detail, never added into any token sum, or the total would
-    // double-count exactly the reasoning share.
-    rec.reasoningOutput = Number(lastUsage.reasoning_output_tokens) || 0;
-  }
-
-  rec.title = maskSecrets(clip(firstPrompt)) || '(untitled)';
-  return { session: seal(rec), turns, parseStats: stats };
-}
 
 // ── file discovery ──────────────────────────────────────────────────────────
 
@@ -1005,236 +348,6 @@ async function loadDeps(override) {
     };
   }
   return _deps;
-}
-
-// ── aggregation ─────────────────────────────────────────────────────────────
-
-// Buckets are keyed by transcript-derived strings (model ids, project names), so
-// a session naming itself `__proto__` would hit Object.prototype's setter: the
-// bucket never becomes an own property and vanishes from JSON.stringify, silently
-// losing that model's spend. Callers build these maps with Object.create(null);
-// this guard keeps the invariant local even if one forgets.
-function bucket(map, key) {
-  if (!Object.prototype.hasOwnProperty.call(map, key)) {
-    map[key] = {
-      sessions: 0, responses: 0, input: 0, output: 0, cacheRead: 0, cacheWrite: 0,
-      tokens: 0, cost: 0,
-      // `minutes` and `confidence` are NOT decoration. Three consumers read them
-      // and every one of them silently rendered a zero before they existed:
-      // project rows showed "0m", every category showed "confidence 0.00" (which
-      // ADR-0009 §5 requires be DISPLAYED — a constant 0.00 is worse than
-      // omitting it), and high-volume-automation printed "0.0 minutes each"
-      // while its evidence asserted "short duration" — a duration claim it never
-      // measured, i.e. fabricated evidence under ADR-0009 §6 rule 1.
-      minutes: 0,
-      confidence: 0,   // session-weighted mean, finalised by sealBuckets()
-    };
-  }
-  return map[key];
-}
-
-function addTo(b, s) {
-  b.sessions++; b.responses += s.responses;
-  b.input += s.input; b.output += s.output;
-  b.cacheRead += s.cacheRead; b.cacheWrite += s.cacheWrite;
-  b.tokens += s.tokens; b.cost += s.cost;
-  b.minutes += Number(s.minutes) || 0;
-  // Accumulated as a SUM here and divided by `sessions` in sealBuckets, so the
-  // result is a true mean rather than a running average that drifts with order.
-  b.confidence += Number(s.confidence) || 0;
-}
-
-/** Turn accumulated confidence sums into means. Must run after every addTo. */
-function sealBuckets(...maps) {
-  for (const map of maps) {
-    for (const k of Object.keys(map)) {
-      const b = map[k];
-      b.confidence = b.sessions ? round(b.confidence / b.sessions, 3) : 0;
-      b.minutes = round(b.minutes, 2);
-    }
-  }
-}
-
-const round = (n, p = 6) => Math.round(n * 10 ** p) / 10 ** p;
-
-/** Turn cached per-file records into the Aggregate the UI and detectors read. */
-function aggregate(records, { days, now, cutoff, deps }) {
-  const sessions = [];
-  // Null-prototype: these are keyed by transcript-derived strings (day, model id,
-  // provider, project, category), so `__proto__` as a key must be an ordinary
-  // bucket, not a prototype write that silently discards the data.
-  const byDay = Object.create(null);
-  const byModel = Object.create(null);
-
-  for (const rec of records) {
-    if (!rec || !rec.responses) continue;                 // no assistant turn → not a session
-    if (rec.end === null || rec.end < cutoff) continue;    // outside the window
-
-    let input = 0, output = 0, cacheRead = 0, cacheWrite = 0, cost = 0;
-    let firstDay = null;
-    const activeDays = new Set();
-    for (const row of rec.usage) {
-      if (firstDay === null || row.day < firstDay) firstDay = row.day;
-      // `day` prices the row at the rate in effect WHEN THOSE TOKENS WERE
-      // SPENT, not today's. A published rate change (Sonnet 5's introductory
-      // period ending 2026-09-01) must not retroactively restate a finished
-      // window — August's spend was metered at August's rate and has to keep
-      // reading that way. Rows are already keyed by day, so this costs nothing.
-      // costObserved (opencode): the transcript's OWN metered figure for the
-      // row, when present, outranks the pricing table — observed truth beats a
-      // rate ak must guess (kimi/openrouter/local). null means no observation
-      // and the table applies, never a fabricated $0.
-      const rowCost = row.costObserved != null ? row.costObserved : (deps.costOf({
-        model: row.model, provider: rec.provider, day: row.day,
-        input: row.input, output: row.output, cacheRead: row.cacheRead, cacheWrite: row.cacheWrite,
-      }) || 0);
-      input += row.input; output += row.output;
-      cacheRead += row.cacheRead; cacheWrite += row.cacheWrite;
-      cost += rowCost;
-
-      const rowTokens = row.input + row.output + row.cacheRead + row.cacheWrite;
-      if (!byDay[row.day]) byDay[row.day] = { tokens: 0, cost: 0, sessions: 0, sessionsActive: 0 };
-      byDay[row.day].tokens += rowTokens;
-      byDay[row.day].cost = round(byDay[row.day].cost + rowCost);
-      activeDays.add(row.day);
-      const m = bucket(byModel, row.model);
-      m.responses += row.responses; m.input += row.input; m.output += row.output;
-      m.cacheRead += row.cacheRead; m.cacheWrite += row.cacheWrite;
-      m.tokens += rowTokens; m.cost = round(m.cost + rowCost);
-    }
-    for (const day of activeDays) byDay[day].sessionsActive++;
-
-    const verdict = deps.classify({
-      title: rec.title, skill: rec.skill, plugin: rec.plugin,
-      tools: rec.tools, prompts: rec.prompts, responses: rec.responses,
-    }) ?? {};
-
-    sessions.push({
-      id: rec.id, host: rec.host ?? rec.provider,
-      provider: rec.inferenceProvider ?? null,
-      transcriptProvider: rec.provider,
-      providerProvenance: rec.providerProvenance ?? 'unknown',
-      title: rec.title, project: rec.project,
-      worktree: rec.worktree ?? null,
-      start: new Date(rec.start ?? rec.end).toISOString(),
-      minutes: Math.round(((rec.end - (rec.start ?? rec.end)) / 60_000) * 10) / 10,
-      prompts: rec.prompts, responses: rec.responses, exceptions: rec.exceptions,
-      sidechain: rec.sidechain, threadSource: rec.threadSource,
-      models: rec.models.slice(),
-      input, output, cacheRead, cacheWrite,
-      tokens: input + output + cacheRead + cacheWrite,
-      cost: round(cost),
-      tools: { ...rec.tools },
-      category: verdict.category ?? 'Unclassified',
-      confidence: verdict.confidence ?? 0,
-      basis: verdict.basis ?? 'no signal',
-      skill: rec.skill, plugin: rec.plugin,
-      // Codex-only detail (v6); zero / null on Claude sessions and on v5-cached
-      // records (the schema bump re-derives those).
-      reasoningOutput: rec.reasoningOutput ?? 0,
-      rateLimits: rec.rateLimits ?? null,
-      _span: [rec.start ?? rec.end, rec.end],
-      // Pre-v2 cache entries have no `active`; fall back to the whole span so a
-      // stale record degrades to the old figure instead of vanishing.
-      _active: Array.isArray(rec.active) && rec.active.length ? rec.active : [[rec.start ?? rec.end, rec.end]],
-      _punchcard: rec.punchcard,
-      // The day this session's tokens FIRST landed on — always a key of byDay,
-      // which keeps sum(byDay.sessions) === totals.sessions. Its start day is
-      // not usable: a session can open at 23:58 and only bill after midnight.
-      _day: firstDay,
-    });
-  }
-
-  sessions.sort((a, b) => b.cost - a.cost || Date.parse(b.start) - Date.parse(a.start));
-
-  const totals = {
-    sessions: sessions.length, responses: 0, exceptions: 0, input: 0, output: 0,
-    cacheRead: 0, cacheWrite: 0, tokens: 0, cost: 0,
-    spanMinutes: 0, spanUnionSeconds: 0, engagedSeconds: 0,
-  };
-  const byHost = Object.create(null), byProvider = Object.create(null);
-  const byProject = Object.create(null);
-  const byCategory = Object.create(null), punchcard = Object.create(null);
-  const tree = new Map();
-  let spanMs = 0;
-
-  for (const s of sessions) {
-    totals.responses += s.responses; totals.exceptions += s.exceptions;
-    totals.input += s.input; totals.output += s.output;
-    totals.cacheRead += s.cacheRead; totals.cacheWrite += s.cacheWrite;
-    totals.tokens += s.tokens; totals.cost += s.cost;
-    spanMs += s._span[1] - s._span[0];
-
-    addTo(bucket(byHost, s.host ?? 'unknown'), s);
-    addTo(bucket(byProvider, s.provider ?? 'unknown'), s);
-    addTo(bucket(byProject, s.project), s);
-    addTo(bucket(byCategory, s.category), s);
-    // A session that used two models counts once under EACH — the token and
-    // cost columns already partition cleanly, session counts cannot.
-    // byModel previously got a bare sessions++ while the other three ran addTo,
-    // so it was the same TYPE with different FIELDS filled — the shape looked
-    // uniform and was not. Accumulate minutes/confidence here too.
-    for (const model of s.models) {
-      const b = bucket(byModel, model);
-      b.sessions++;
-      b.minutes += Number(s.minutes) || 0;
-      b.confidence += Number(s.confidence) || 0;
-    }
-    if (s._day && byDay[s._day]) byDay[s._day].sessions++;
-    for (const [k, n] of Object.entries(s._punchcard)) punchcard[k] = (punchcard[k] ?? 0) + n;
-
-    if (!tree.has(s.project)) tree.set(s.project, { project: s.project, sessions: 0, cost: 0, tokens: 0, minutes: 0, cats: new Map(), rows: [] });
-    const node = tree.get(s.project);
-    node.sessions++; node.cost = round(node.cost + s.cost); node.tokens += s.tokens;
-    node.minutes = Math.round((node.minutes + s.minutes) * 10) / 10;
-    const cat = node.cats.get(s.category) ?? { category: s.category, sessions: 0, cost: 0 };
-    cat.sessions++; cat.cost = round(cat.cost + s.cost);
-    node.cats.set(s.category, cat);
-    node.rows.push(s);
-  }
-
-  totals.cost = round(totals.cost);
-  // Three tiers, each honest about a different thing:
-  //   engagedSeconds   — union of ACTIVE intervals: time actually worked
-  //   spanUnionSeconds — union of whole spans: wall-clock with a session open
-  //   spanMinutes      — sum of spans: the double-counting figure, kept as the
-  //                      clearly-labelled secondary the ADR asks the UI to show
-  sealBuckets(byHost, byProvider, byProject, byCategory, byModel);
-
-  totals.spanMinutes = Math.round((spanMs / 60_000) * 10) / 10;
-  totals.spanUnionSeconds = mergeIntervals(sessions.map((s) => s._span));
-  totals.engagedSeconds = mergeIntervals(sessions.flatMap((s) => s._active));
-
-  const projectTree = [...tree.values()]
-    .map((n) => ({
-      project: n.project, sessions: n.sessions, cost: n.cost, tokens: n.tokens, minutes: n.minutes,
-      categories: [...n.cats.values()].sort((a, b) => b.cost - a.cost || b.sessions - a.sessions),
-      rows: n.rows,
-    }))
-    .sort((a, b) => b.cost - a.cost || b.tokens - a.tokens);
-
-  for (const s of sessions) { delete s._span; delete s._active; delete s._punchcard; delete s._day; }
-
-  // Local rate-limit history: every Codex rollout embeds live quota snapshots
-  // in its token_count events, so a utilization time series is reconstructable
-  // retroactively with ZERO network — one point per session (its last
-  // snapshot), oldest first. Claude has no local analogue (its quota arrives
-  // via the statusline push — see quota.mjs), hence codex-prefixed.
-  const codexRateLimits = sessions
-    .filter((s) => s.host === 'codex' && s.rateLimits && Number.isFinite(s.rateLimits.at))
-    .map((s) => s.rateLimits)
-    .sort((x, y) => x.at - y.at);
-
-  const agg = {
-    generatedAt: new Date(now).toISOString(),
-    windowDays: days,
-    pricesAsOf: deps.pricesAsOf ?? null,
-    totals, byDay, byModel, byHost, byProvider,
-    byProject, byCategory,
-    punchcard, projectTree, sessions, codexRateLimits, insights: [],
-  };
-  agg.insights = deps.detectInsights(agg) ?? [];
-  return agg;
 }
 
 // ── build ───────────────────────────────────────────────────────────────────
@@ -1517,33 +630,6 @@ async function scan(o = {}) {
 }
 
 /**
- * Overlay Codex's SQLite thread ledger onto parsed session records. Pure —
- * returns copies where anything changes, never mutates a (possibly cached)
- * record. Two corrections, both attribution-only:
- *   - a session whose rollout carried no `thread_source` is backfilled from
- *     the ledger's `threads.thread_source` (or marked `subagent` when a
- *     spawn edge names it as a child);
- *   - a session the ledger says is a subagent has its token usage STRIPPED,
- *     mirroring the parse-time exclusion: its rollout replays the parent's
- *     entire token history, so keeping the tokens double-counts the parent
- *     (ccusage/ccusage#950). The record itself stays visible/auditable.
- * Exported for test.
- */
-export function applyCodexLedger(records, ledger) {
-  if (!ledger || !(ledger.threads instanceof Map)) return records;
-  return records.map((rec) => {
-    if (!rec || rec.provider !== 'codex') return rec;
-    const t = ledger.threads.get(rec.id);
-    const fromEdges = ledger.parents instanceof Map && ledger.parents.has(rec.id) ? 'subagent' : null;
-    const source = rec.threadSource ?? t?.threadSource ?? fromEdges;
-    if (source === rec.threadSource && (source !== 'subagent' || !rec.usage.length)) return rec;
-    const out = { ...rec, threadSource: source };
-    if (source === 'subagent' && out.usage.length) out.usage = [];
-    return out;
-  });
-}
-
-/**
  * The Aggregate, memoized in-process and refreshed when stale. This is the
  * dashboard's read path — `maxAgeMs` (default 15 s) bounds how often a poll can
  * trigger a re-stat of the corpus.
@@ -1682,55 +768,4 @@ export async function readSession(id, o = {}) {
   } catch { return null; }
 
   return sessionPayload(parsed.session, parsed.turns, await loadDeps(o.deps));
-}
-
-/** The /api/session payload for any parsed record (claude, codex, opencode):
- *  meta with pricer-backed cost, and secret-masked, truncation-signalled turns. */
-function sessionPayload(rec, turns, deps) {
-  const usage = (rec.usage ?? []).reduce((a, row) => ({
-    input: a.input + row.input, output: a.output + row.output,
-    cacheRead: a.cacheRead + row.cacheRead, cacheWrite: a.cacheWrite + row.cacheWrite,
-  }), { input: 0, output: 0, cacheRead: 0, cacheWrite: 0 });
-
-  return {
-    meta: {
-      id: rec.id, host: rec.host ?? rec.provider,
-      provider: rec.inferenceProvider ?? null,
-      transcriptProvider: rec.provider,
-      providerProvenance: rec.providerProvenance ?? 'unknown',
-      title: rec.title, project: rec.project,
-      worktree: rec.worktree ?? null,
-      start: rec.start === null ? null : new Date(rec.start).toISOString(),
-      end: rec.end === null ? null : new Date(rec.end).toISOString(),
-      minutes: rec.start === null ? 0 : Math.round(((rec.end - rec.start) / 60_000) * 10) / 10,
-      prompts: rec.prompts, responses: rec.responses, exceptions: rec.exceptions,
-      sidechain: rec.sidechain, threadSource: rec.threadSource,
-      models: rec.models.slice(), tools: { ...rec.tools },
-      skill: rec.skill, plugin: rec.plugin,
-      // Priced here from the same per-model rows aggregate() uses, rather than
-      // left undefined: the transcript header rendered a hardcoded "$0.00" on a
-      // panel whose whole subject is cost. `.filter(Boolean)` could not drop it
-      // because fmtUsd(undefined) is the truthy string "$0.00".
-      cost: sessionCost(rec, deps),
-      ...usage, tokens: usage.input + usage.output + usage.cacheRead + usage.cacheWrite,
-    },
-    // ADR-0009 §8: truncation is the other way content is withheld, and it used
-    // to be silent — `truncated` was set and the renderer ignored it, so an
-    // abridged turn read as a complete one. Both keys are emitted only when the
-    // slice actually fired, so the field's *presence* is the signal and a whole
-    // turn cannot be misread as an abridged one. `originalChars` is measured
-    // after `maskSecrets`, so it describes loss due to truncation alone — it is
-    // not a raw-file length, and must not be rendered as one.
-    turns: (turns ?? []).map((t) => {
-      const text = maskSecrets(t.text);
-      const originalChars = text.length;
-      if (originalChars <= MAX_TURN_CHARS) return { ...t, text };
-      return {
-        ...t,
-        text: `${text.slice(0, MAX_TURN_CHARS)}\n…[truncated]`,
-        truncated: true,
-        originalChars,
-      };
-    }),
-  };
 }
